+++ src/metadata/mod.rs
@@ -16,10 +16,7 @@ mod writing;
 
 use self::adult_reader::AdultReader;
 use super::node_ops::NodeDuty;
-use crate::{
-    capacity::ChunkHolderDbs, chunk_store::UsedSpace, network::Network, node_ops::NodeDuties,
-    Result,
-};
+use crate::{capacity::ChunkHolderDbs, chunk_store::UsedSpace, node_ops::NodeDuties, Result};
 use blob_register::BlobRegister;
 use elder_stores::ElderStores;
 use map_storage::MapStorage;