+++ src/node/mod.rs
@@ -13,31 +13,23 @@ mod messaging;
 mod split;
 
 use crate::{
-    capacity::{Capacity, ChunkHolderDbs, RateLimit},
     chunk_store::UsedSpace,
     chunks::Chunks,
     event_mapping::{map_routing_event, LazyError, Mapping, MsgContext},
-    metadata::{adult_reader::AdultReader, Metadata},
+    metadata::Metadata,
     network::Network,
-    node_ops::{NodeDuties, NodeDuty},
+    node_ops::NodeDuty,
     section_funds::SectionFunds,
-    state_db::store_new_reward_keypair,
-    transfers::{get_replicas::transfer_replicas, Transfers},
+    state_db::{get_reward_pk, store_new_reward_keypair},
+    transfers::Transfers,
     Config, Error, Result,
 };
 use bls::SecretKey;
-use ed25519_dalek::PublicKey as Ed25519PublicKey;
-use futures::lock::Mutex;
-use hex_fmt::HexFmt;
-use log::{debug, error, info, trace, warn};
-use sn_data_types::{ActorHistory, PublicKey, TransferPropagated, WalletHistory};
-use sn_messaging::{client::Message, DstLocation, SrcLocation};
-use sn_routing::{Event as RoutingEvent, EventStream, NodeElderChange, MIN_AGE};
-use sn_routing::{Prefix, XorName, ELDER_SIZE as GENESIS_ELDER_COUNT};
-use sn_transfers::{TransferActor, Wallet};
-use std::collections::BTreeMap;
+use log::{error, info};
+use sn_data_types::PublicKey;
+use sn_routing::EventStream;
+use sn_routing::{Prefix, XorName};
 use std::path::{Path, PathBuf};
-use std::sync::Arc;
 use std::{
     fmt::{self, Display, Formatter},
     net::SocketAddr,
@@ -47,13 +39,7 @@ use std::{
 #[derive(Clone)]
 pub struct NodeInfo {
     ///
-    pub genesis: bool,
-    ///
     pub root_dir: PathBuf,
-    ///
-    pub node_name: XorName,
-    ///
-    pub node_id: Ed25519PublicKey,
     /// The key used by the node to receive earned rewards.
     pub reward_key: PublicKey,
 }
@@ -88,28 +74,28 @@ enum Role {
 impl Role {
     fn as_adult(&self) -> Result<&AdultRole> {
         match self {
-            Self::Adult(adult_state) => Ok(adult_state),
+            Self::Adult(adult) => Ok(adult),
             _ => Err(Error::NotAnAdult),
         }
     }
 
     fn as_adult_mut(&mut self) -> Result<&mut AdultRole> {
         match self {
-            Self::Adult(adult_state) => Ok(adult_state),
+            Self::Adult(adult) => Ok(adult),
             _ => Err(Error::NotAnAdult),
         }
     }
 
     fn as_elder(&self) -> Result<&ElderRole> {
         match self {
-            Self::Elder(elder_state) => Ok(elder_state),
+            Self::Elder(elder) => Ok(elder),
             _ => Err(Error::NotAnElder),
         }
     }
 
     fn as_elder_mut(&mut self) -> Result<&mut ElderRole> {
         match self {
-            Self::Elder(elder_state) => Ok(elder_state),
+            Self::Elder(elder) => Ok(elder),
             _ => Err(Error::NotAnElder),
         }
     }
@@ -121,7 +107,6 @@ pub struct Node {
     network_events: EventStream,
     node_info: NodeInfo,
     used_space: UsedSpace,
-    prefix: Prefix,
     role: Role,
 }
 
@@ -135,45 +120,28 @@ impl Node {
         let root_dir = root_dir_buf.as_path();
         std::fs::create_dir_all(root_dir)?;
 
-        let reward_key_task = async move {
-            let res: Result<PublicKey>;
-            match config.wallet_id() {
-                Some(public_key) => {
-                    res = Ok(PublicKey::Bls(crate::state_db::pk_from_hex(public_key)?));
-                }
-                None => {
-                    let secret = SecretKey::random();
-                    let public = secret.public_key();
-                    store_new_reward_keypair(root_dir, &secret, &public).await?;
-                    res = Ok(PublicKey::Bls(public));
-                }
-            };
-            res
-        }
-        .await;
+        let reward_key = match get_reward_pk(root_dir).await? {
+            Some(public_key) => PublicKey::Bls(public_key),
+            None => {
+                let secret = SecretKey::random();
+                let public = secret.public_key();
+                store_new_reward_keypair(root_dir, &secret, &public).await?;
+                PublicKey::Bls(public)
+            }
+        };
 
-        let reward_key = reward_key_task?;
-        let (network_api, network_events) = Network::new(config).await?;
+        let (network_api, network_events) = Network::new(root_dir, config).await?;
 
         let node_info = NodeInfo {
-            genesis: config.is_first(),
             root_dir: root_dir_buf,
-            node_name: network_api.our_name().await,
-            node_id: network_api.public_key().await,
             reward_key,
         };
 
         let used_space = UsedSpace::new(config.max_capacity());
 
         let node = Self {
-            prefix: network_api.our_prefix().await,
             role: Role::Adult(AdultRole {
-                chunks: Chunks::new(
-                    node_info.node_name,
-                    node_info.root_dir.as_path(),
-                    used_space.clone(),
-                )
-                .await?,
+                chunks: Chunks::new(node_info.root_dir.as_path(), used_space.clone()).await?,
             }),
             node_info,
             used_space,
@@ -181,7 +149,7 @@ impl Node {
             network_events,
         };
 
-        messaging::send(node.register_wallet().await, &node.network_api).await;
+        messaging::send(node.register_wallet().await, &node.network_api).await?;
 
         Ok(node)
     }
@@ -248,7 +216,7 @@ fn handle_error(err: LazyError) {
 fn try_handle_error(err: Error, ctx: Option<MsgContext>) {
     use std::error::Error;
     if let Some(source) = err.source() {
-        if let Some(ctx) = ctx {
+        if let Some(_ctx) = ctx {
             info!(
                 "unimplemented: Handle errors. This should be return w/ lazyError to sender. {:?}",
                 err